simplify stories
